+++ ICD.Profound.ConnectPRO/Themes/UserInterface/Presenters/Common/Cameras/CameraControlPresenter.cs
@@ -196,6 +196,12 @@ namespace ICD.Profound.ConnectPRO.Themes.UserInterface.Presenters.Common.Cameras
 				m_SelectedCamera.Tilt(action);
 		}
 
+		private void Home()
+		{
+			if (m_SelectedCamera != null && m_SelectedCamera.SupportedCameraFeatures.HasFlag(eCameraFeatures.Home))
+				m_SelectedCamera.SendCameraHome();
+		}
+
 		#endregion
 
 		#region Camera Callbacks
@@ -251,6 +257,7 @@ namespace ICD.Profound.ConnectPRO.Themes.UserInterface.Presenters.Common.Cameras
 			view.OnCameraMoveLeftButtonPressed += ViewOnCameraMoveLeftButtonPressed;
 			view.OnCameraMoveRightButtonPressed += ViewOnCameraMoveRightButtonPressed;
 			view.OnCameraMoveUpButtonPressed += ViewOnCameraMoveUpButtonPressed;
+			view.OnCameraHomeButtonPressed += ViewOnCameraHomeButtonPressed;
 			view.OnCameraZoomInButtonPressed += ViewOnCameraZoomInButtonPressed;
 			view.OnCameraZoomOutButtonPressed += ViewOnCameraZoomOutButtonPressed;
 			view.OnPresetButtonReleased += ViewOnPresetButtonReleased;
@@ -271,6 +278,7 @@ namespace ICD.Profound.ConnectPRO.Themes.UserInterface.Presenters.Common.Cameras
 			view.OnCameraMoveLeftButtonPressed -= ViewOnCameraMoveLeftButtonPressed;
 			view.OnCameraMoveRightButtonPressed -= ViewOnCameraMoveRightButtonPressed;
 			view.OnCameraMoveUpButtonPressed -= ViewOnCameraMoveUpButtonPressed;
+			view.OnCameraHomeButtonPressed -= ViewOnCameraHomeButtonPressed;
 			view.OnCameraZoomInButtonPressed -= ViewOnCameraZoomInButtonPressed;
 			view.OnCameraZoomOutButtonPressed -= ViewOnCameraZoomOutButtonPressed;
 			view.OnPresetButtonReleased -= ViewOnPresetButtonReleased;
@@ -305,6 +313,11 @@ namespace ICD.Profound.ConnectPRO.Themes.UserInterface.Presenters.Common.Cameras
 		private void ViewOnCameraMoveUpButtonPressed(object sender, EventArgs eventArgs)
 		{
 			Tilt(eCameraTiltAction.Up);
+		}
+
+		private void ViewOnCameraHomeButtonPressed(object sender, EventArgs eventArgs)
+		{
+			Home();
 		}
 
 		private void ViewOnCameraZoomInButtonPressed(object sender, EventArgs eventArgs)